move build output parsing from task dll to vs package
