+++ VSRAD.BuildTools/VSRAD.BuildTools.csproj
@@ -48,12 +48,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Errors\LineMapper.cs" />
-    <Compile Include="Errors\Message.cs" />
     <Compile Include="IPCBuildResult.cs" />
     <Compile Include="RemoteBuildTask.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Errors\Parser.cs" />
     <Compile Include="IPCBridge.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
