+++ VSRAD.Package/VSRAD.Package.csproj
@@ -122,6 +122,8 @@
     <Compile Include="..\VSRAD.DebugServer\IPC\StreamExtensions.cs">
       <Link>Server\IPC\StreamExtensions.cs</Link>
     </Compile>
+    <Compile Include="BuildTools\Errors\LineMapper.cs" />
+    <Compile Include="BuildTools\Errors\Parser.cs" />
     <Compile Include="Colors.cs" />
     <Compile Include="Commands\AddToWatchesCommand.cs" />
     <Compile Include="Commands\AddArrayToWatchesCommand.cs" />
