+++ VSRAD.PackageTests/VSRAD.PackageTests.csproj
@@ -61,6 +61,8 @@
   </Choose>
   <ItemGroup>
     <Compile Include="BuildTools\BuildToolsServerTest.cs" />
+    <Compile Include="BuildTools\Errors\LineMapperTests.cs" />
+    <Compile Include="BuildTools\Errors\ParserTests.cs" />
     <Compile Include="DebugVisualizer\ColumnSelectorTests.cs" />
     <Compile Include="DebugVisualizer\ColumnStylingTests.cs" />
     <Compile Include="DebugVisualizer\GroupIndexSelectorTests.cs" />
